add group_by operator
