+++ src/json/mod.rs
@@ -1,6 +1,9 @@
 mod async_value;
+
+mod group_by;
 mod json_like;
 mod json_schema;
 mod serde_value;
+pub use group_by::*;
 pub use json_like::*;
 pub use json_schema::*;
