+++ src/json/serde_value.rs
@@ -39,19 +39,11 @@ impl JsonLike for serde_json::Value {
         Some(val)
     }
 
-    fn from_output(value: &Self::Output) -> &Self {
-        value
-    }
-
     fn as_string_ok(&self) -> Option<&String> {
         match self {
             serde_json::Value::String(s) => Some(s),
             _ => None,
         }
-    }
-
-    fn to_output(value: &Self) -> &Self::Output {
-        value
     }
 }
